Add Put and Delete decorators
