+++ mod.ts
@@ -8,6 +8,9 @@ export * from "./drivers/core_adapter.ts";
 export * from "./registration.ts";
 export * from "./route_decorators.ts";
 export * from "./post_decorator.ts";
+export * from "./get_decorator.ts";
+export * from "./put_decorator.ts";
+export * from "./delete_decorator.ts";
 export * from "./model_decorators.ts";
 export * from "./model_builders.ts";
 export * from "./context.ts";
